+++ src/Model/Table/UsersTable.php
@@ -116,18 +116,6 @@ class UsersTable extends AppTable
     }
 
     /**
-     * Gets the users list
-     * @return array List
-     * @uses $cache
-     */
-    public function getList()
-    {
-        return $this->find('list')
-            ->cache('users_list', $this->cache)
-            ->toArray();
-    }
-
-    /**
      * Initialize method
      * @param array $config The configuration for the table
      * @return void