+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## 2.14 branch
 ### 2.14.12
 * now the `PhotosAlbum` entity has the `path` virtual field;
+* `getList()` and `getTreeList()` methods moved to `AppModel` class. So now
+    each model has these methods;
 * fixed a serious bug in the association of tokens with users;
 * fixed a serious bug for the `AppTable::isOwnedBy()` method;
 * fixed bug for validator for pages categories;